+++ packages/algob/src/lib/compile.ts
@@ -1,6 +1,6 @@
 import { getPathFromDirRecursive } from "@algo-builder/runtime";
 import { BuilderError, ERRORS, parseAlgorandError } from "@algo-builder/web";
-import type { Algodv2, CompileOut } from "algosdk";
+import type { Algodv2, modelsv2 } from "algosdk";
 import { spawnSync, SpawnSyncReturns } from "child_process";
 import * as fs from 'fs';
 import * as murmurhash from 'murmurhash';
@@ -86,7 +86,7 @@ export class CompileOp {
     }
   }
 
-  callCompiler (code: string): Promise<CompileOut> {
+  callCompiler (code: string): Promise<modelsv2.CompileResponse> {
     return this.algocl.compile(code).do();
   }